pyproject.toml
[tool.poetry]
name = "orisa_api"
version = "0.1.0"
description = ""
authors = ["vzagoruyko <vitalik.zgrk@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
django-nested-admin = "^4.0.2"
docxtpl = "^0.16.5"
python-docx = "^0.8.11"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
huey = "^2.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

